Added VersionHistoryIT integration test to validate _valid_from and _valid_to columns.
