+++ sink-connector-lightweight/src/test/java/com/altinity/clickhouse/debezium/embedded/history/BinLogHistoryIT.java
@@ -259,176 +259,4 @@ public class BinLogHistoryIT {
         
         log.info("Successfully validated {} temporal tracking columns for {}.{}", columnCount, database, table);
     }
-    
-    /**
-     * Test that validates _valid_to and _valid_from columns are correctly set for
-     * INSERT, UPDATE, and DELETE operations.
-     * 
-     * For INSERT: _valid_to should be set to max date (2100-01-01)
-     * For UPDATE: 
-     *   - Old record should be closed (is_deleted=1, _valid_to set to binlog timestamp)
-     *   - New record should have open-ended _valid_to (max date)
-     * For DELETE: record should be marked as deleted (is_deleted=1, _valid_to set to binlog timestamp)
-     */
-    @DisplayName("Test that validates _valid_to and _valid_from columns for INSERT, UPDATE, DELETE operations")
-    @Test
-    public void testValidToValidFromColumnsOnUpdateDelete() throws Exception {
-        
-        Injector injector = Guice.createInjector(new AppInjector());
-
-        Properties props = getDebeziumProperties(mySqlContainer, clickHouseContainer);
-        props.setProperty("snapshot.mode", "no_data");
-        props.setProperty("schema.history.internal.store.only.captured.tables.ddl", "true");
-        props.setProperty("schema.history.internal.store.only.captured.databases.ddl", "true");
-        props.setProperty("database.include.list", "employees");
-
-        // Enable binlog history
-        props.setProperty("replication.history.enable", "true");
-
-        ClickHouseDebeziumEmbeddedApplication clickHouseDebeziumEmbeddedApplication = new ClickHouseDebeziumEmbeddedApplication();
-
-        ExecutorService executorService = Executors.newFixedThreadPool(1);
-        executorService.execute(() -> {
-            try {
-                clickHouseDebeziumEmbeddedApplication.start(injector.getInstance(DebeziumRecordParserService.class), props, false);
-                DebeziumEmbeddedRestApi.startRestApi(props, injector, clickHouseDebeziumEmbeddedApplication.getDebeziumEventCapture(), new Properties());
-            } catch (Exception e) {
-                throw new RuntimeException(e);
-            }
-        });
-
-        Thread.sleep(25000);
-
-        // Step 1: Create table and insert a record
-        Connection conn = ITCommon.connectToMySQL(mySqlContainer);
-        conn.prepareStatement("create table `employees_temporal_test`(emp_id int not null, name varchar(255), salary int, primary key(emp_id))").execute();
-        
-        // Insert first record
-        conn.prepareStatement("insert into employees_temporal_test values(1, 'John Doe', 50000)").execute();
-        
-        Thread.sleep(10000);
-        
-        BaseDbWriter writer = ITCommon.getDBWriter(clickHouseContainer);
-        
-        // Validate INSERT: _valid_to should be max date (2100-01-01)
-        log.info("Validating INSERT operation - _valid_to should be max date");
-        ResultSet insertRs = ITCommon.executeQueryWithResultSet(
-            "SELECT emp_id, name, salary, `_valid_to`, `is_deleted`, `_version` FROM employees.employees_temporal_test ORDER BY `_version`",
-            writer.getConnection());
-        
-        int insertRecordCount = 0;
-        while (insertRs.next()) {
-            insertRecordCount++;
-            String validTo = insertRs.getString("_valid_to");
-            int isDeleted = insertRs.getInt("is_deleted");
-            log.info("After INSERT - Record: emp_id={}, name={}, _valid_to={}, is_deleted={}", 
-                insertRs.getInt("emp_id"), insertRs.getString("name"), validTo, isDeleted);
-            
-            // For inserted record, _valid_to should be max date and is_deleted should be 0
-            assertTrue("After INSERT, is_deleted should be 0", isDeleted == 0);
-            assertTrue("After INSERT, _valid_to should be year 2100", validTo.startsWith("2100"));
-        }
-        assertTrue("Should have exactly 1 record after INSERT", insertRecordCount == 1);
-        
-        // Step 2: Update the record
-        log.info("Performing UPDATE operation");
-        conn.prepareStatement("update employees_temporal_test set salary = 60000 where emp_id = 1").execute();
-        
-        Thread.sleep(10000);
-        
-        // Validate UPDATE: Should have multiple records - old one closed, new one with open _valid_to
-        log.info("Validating UPDATE operation - checking _valid_to values");
-        ResultSet updateRs = ITCommon.executeQueryWithResultSet(
-            "SELECT emp_id, name, salary, `_valid_to`, `is_deleted`, `_version` FROM employees.employees_temporal_test ORDER BY `_version`",
-            writer.getConnection());
-        
-        int updateRecordCount = 0;
-        int activeRecords = 0;
-        int closedRecords = 0;
-        int latestSalary = 0;
-        
-        while (updateRs.next()) {
-            updateRecordCount++;
-            String validTo = updateRs.getString("_valid_to");
-            int isDeleted = updateRs.getInt("is_deleted");
-            int salary = updateRs.getInt("salary");
-            log.info("After UPDATE - Record {}: emp_id={}, salary={}, _valid_to={}, is_deleted={}", 
-                updateRecordCount, updateRs.getInt("emp_id"), salary, validTo, isDeleted);
-            
-            if (isDeleted == 0 && validTo.startsWith("2100")) {
-                activeRecords++;
-                latestSalary = salary;
-            } else {
-                closedRecords++;
-            }
-        }
-        
-        log.info("After UPDATE - Total records: {}, Active records: {}, Closed records: {}", 
-            updateRecordCount, activeRecords, closedRecords);
-        
-        // Validate that we have at least one active record with updated salary
-        assertTrue("Should have at least 1 active record after UPDATE", activeRecords >= 1);
-        assertTrue("Active record should have updated salary of 60000", latestSalary == 60000);
-        
-        // Validate using FINAL that we get the latest version with updated salary
-        ResultSet finalRs = ITCommon.executeQueryWithResultSet(
-            "SELECT emp_id, name, salary, `_valid_to`, `is_deleted` FROM employees.employees_temporal_test FINAL WHERE is_deleted = 0",
-            writer.getConnection());
-        
-        boolean foundActiveRecord = false;
-        while (finalRs.next()) {
-            foundActiveRecord = true;
-            int salary = finalRs.getInt("salary");
-            String validTo = finalRs.getString("_valid_to");
-            int isDeleted = finalRs.getInt("is_deleted");
-            log.info("FINAL after UPDATE - emp_id={}, salary={}, _valid_to={}, is_deleted={}", 
-                finalRs.getInt("emp_id"), salary, validTo, isDeleted);
-            
-            assertTrue("FINAL record should have updated salary of 60000", salary == 60000);
-            assertTrue("FINAL record should have is_deleted = 0", isDeleted == 0);
-            assertTrue("FINAL record should have _valid_to in 2100", validTo.startsWith("2100"));
-        }
-        assertTrue("Should find active record with FINAL after UPDATE", foundActiveRecord);
-        
-        // Step 3: Delete the record
-        log.info("Performing DELETE operation");
-        conn.prepareStatement("delete from employees_temporal_test where emp_id = 1").execute();
-        
-        Thread.sleep(10000);
-        
-        // Validate DELETE: Using FINAL, the record should be marked as deleted
-        log.info("Validating DELETE operation - checking is_deleted and _valid_to");
-        ResultSet deleteRs = ITCommon.executeQueryWithResultSet(
-            "SELECT emp_id, name, salary, `_valid_to`, `is_deleted`, `_version` FROM employees.employees_temporal_test ORDER BY `_version` DESC LIMIT 1",
-            writer.getConnection());
-        
-        while (deleteRs.next()) {
-            int isDeleted = deleteRs.getInt("is_deleted");
-            String validTo = deleteRs.getString("_valid_to");
-            log.info("After DELETE - Latest record: emp_id={}, _valid_to={}, is_deleted={}", 
-                deleteRs.getInt("emp_id"), validTo, isDeleted);
-            
-            assertTrue("After DELETE, is_deleted should be 1 for the latest record", isDeleted == 1);
-            // After delete, _valid_to should NOT be 2100 (it should be the binlog timestamp)
-            assertTrue("After DELETE, _valid_to should not be year 2100 (should be closed)", 
-                !validTo.startsWith("2100"));
-        }
-        
-        // Validate using FINAL that no active records remain
-        ResultSet finalDeleteRs = ITCommon.executeQueryWithResultSet(
-            "SELECT count(*) as cnt FROM employees.employees_temporal_test FINAL WHERE is_deleted = 0",
-            writer.getConnection());
-        
-        while (finalDeleteRs.next()) {
-            int count = finalDeleteRs.getInt("cnt");
-            log.info("FINAL after DELETE - Active record count: {}", count);
-            assertTrue("After DELETE, no active records should remain with FINAL query", count == 0);
-        }
-        
-        log.info("Successfully validated _valid_to and _valid_from columns for INSERT, UPDATE, DELETE operations");
-
-        conn.close();
-        executorService.shutdown();
-        HikariDbSource.close();
-    }
 }
